+++ example/integration_test/in_app_webview/on_download_start_request.dart
@@ -46,7 +46,7 @@ void onDownloadStartRequest() {
 </html>
           """),
           initialSettings:
-              InAppWebViewSettings(clearCache: true, useOnDownloadStart: true),
+              InAppWebViewSettings(clearCache: true),
           onWebViewCreated: (controller) {
             controllerCompleter.complete(controller);
           },