+++ example/integration_test/in_app_webview/on_load_resource.dart
@@ -33,7 +33,7 @@ void onLoadResource() {
             initialFile:
                 "test_assets/in_app_webview_on_load_resource_test.html",
             initialSettings:
-                InAppWebViewSettings(clearCache: true, useOnLoadResource: true),
+                InAppWebViewSettings(clearCache: true),
             onLoadStop: (controller, url) {
               pageLoaded.complete();
             },