+++ example/integration_test/in_app_webview/should_intercept_request.dart
@@ -46,7 +46,7 @@ void shouldInterceptRequest() {
 </html>
                     """),
           initialSettings:
-              InAppWebViewSettings(useShouldInterceptRequest: true),
+              InAppWebViewSettings(),
           onWebViewCreated: (controller) {
             controllerCompleter.complete(controller);
           },