+++ lib/src/print_job/main.dart
@@ -1,2 +1,2 @@
 export 'print_job_controller.dart';
-export 'print_job_settings.dart';
+export 'print_job_settings.dart' show PrintJobSettings;