+++ lib/src/pull_to_refresh/main.dart
@@ -1,2 +1,3 @@
 export 'pull_to_refresh_controller.dart';
-export 'pull_to_refresh_settings.dart';
+export 'pull_to_refresh_settings.dart'
+    show PullToRefreshSettings, PullToRefreshOptions;