+++ lib/src/web_authentication_session/main.dart
@@ -1,2 +1,3 @@
 export 'web_authenticate_session.dart';
-export 'web_authenticate_session_settings.dart';
+export 'web_authenticate_session_settings.dart'
+    show WebAuthenticationSessionSettings;