+++ lib/src/web_authentication_session/web_authenticate_session_settings.dart
@@ -1,8 +1,11 @@
-import 'package:flutter/foundation.dart';
+import 'package:flutter_inappwebview_internal_annotations/flutter_inappwebview_internal_annotations.dart';
 import 'web_authenticate_session.dart';
 
+part 'web_authenticate_session_settings.g.dart';
+
 ///Class that represents the settings that can be used for a [WebAuthenticationSession].
-class WebAuthenticationSessionSettings {
+@ExchangeableObject(copyMethod: true)
+class WebAuthenticationSessionSettings_ {
   ///A Boolean value that indicates whether the session should ask the browser for a private authentication session.
   ///
   ///Set [prefersEphemeralWebBrowserSession] to `true` to request that the browser
@@ -21,38 +24,14 @@ class WebAuthenticationSessionSettings {
   ///**Supported Platforms/Implementations**:
   ///- iOS
   ///- MacOS
-  bool prefersEphemeralWebBrowserSession;
+  bool? prefersEphemeralWebBrowserSession;
 
-  WebAuthenticationSessionSettings(
+  WebAuthenticationSessionSettings_(
       {this.prefersEphemeralWebBrowserSession = false});
 
   Map<String, dynamic> toMap() {
     return {
       "prefersEphemeralWebBrowserSession": prefersEphemeralWebBrowserSession
     };
-  }
-
-  static WebAuthenticationSessionSettings fromMap(Map<String, dynamic> map) {
-    WebAuthenticationSessionSettings settings =
-        new WebAuthenticationSessionSettings();
-    if (defaultTargetPlatform == TargetPlatform.iOS ||
-        defaultTargetPlatform == TargetPlatform.macOS) {
-      settings.prefersEphemeralWebBrowserSession =
-          map["prefersEphemeralWebBrowserSession"];
-    }
-    return settings;
-  }
-
-  Map<String, dynamic> toJson() {
-    return this.toMap();
-  }
-
-  @override
-  String toString() {
-    return toMap().toString();
-  }
-
-  WebAuthenticationSessionSettings copy() {
-    return WebAuthenticationSessionSettings.fromMap(this.toMap());
   }
 }